+++ app/helper/prepare_ml_input.py
@@ -4,6 +4,9 @@ Download and save components from providers, e.g. Hugging Face.
 Components could be models, datasets, tokenizers, metrics etc.
 """
 
+# TODO decorator or facade for handler functions to load provider components,
+# e.g. load_hf_components
+
 from dataclasses import dataclass
 from os import environ as env
 from typing import Final
@@ -29,6 +32,15 @@ from .parse_configs_into_paramdict import ParamDict
 
 @dataclass(repr=False, eq=False)  # slots only >=3.10
 class PipelineOutput:
+    """
+    Holds structured mutable data in the form of
+    - `paramdict` as `ParamDict`
+    - `tokenizer` as `AutoTokenizer`
+    - `dataset_tokenized` as `DatasetDict`
+    - `model`as `AutoModel`
+    - `metrics_loaded` as `list[dict]`
+    """
+
     paramdict: ParamDict
     tokenizer: AutoTokenizer
     dataset_tokenized: DatasetDict
@@ -37,46 +49,99 @@ class PipelineOutput:
 
 
 def prepare_pipeline(paramdict: ParamDict) -> PipelineOutput:
-    """TODO"""
+    """
+    Prepares the pipeline by loading Dataset, Tokenizer, Model and Metrics as well
+    as setting parameter for the used provider in the system environment.\n
+    Expects a populated `ParamDict` and returns a `PipelineOutput`.
+    """
 
     provider = paramdict["sweep"]["provider"]
     _set_provider_env(provider, paramdict[provider])
 
-    dataset_plain = get_dataset_hf(
-        paramdict["dataset"]["dataset"],
-        paramdict["dataset"]["configuration"],
-        paramdict["save_dir"],
-    )
-    num_labels = len(
-        dataset_plain["train"].unique(paramdict["dataset"]["col_to_rename"])
-    )
-    paramdict["dataset"]["num_labels"] = num_labels
-    if debug_on_global:
-        debug(f"The number of unique labels is {num_labels}")
+    return _get_large_components(paramdict)
 
-    tokenizer = get_tokenizer_hf(
-        paramdict["model_full_name"],
-        paramdict["save_dir"],
-    )
 
-    dataset_tokenized = _get_tokenized_dataset(
-        dataset_plain,
-        tokenizer,
-        paramdict["dataset"]["cols_to_tokenize"],
-        paramdict["dataset"]["cols_to_remove"],
-    )
+def _get_large_components(paramdict: ParamDict) -> PipelineOutput:
+    """Loads components needed for the `PipelineOutput`"""
 
-    # model = _get_model(paramdict["model_full_name"],
-    #   paramdict["dataset"]["num_labels"])
-    model = ""
+    # TODO read num_labels from ds card
+    # dataset_plain = _get_dataset(
+    #     paramdict["dataset"]["dataset"],
+    #     paramdict["dataset"]["configuration"],
+    #     paramdict["save_dir"],
+    # )
 
-    metrics_loaded = _get_metrics_to_load_objects(
-        paramdict["metrics"]["metrics_to_load"]
-    )
+    # num_labels = len(
+    #     dataset_plain["train"].unique(paramdict["dataset"]["col_to_rename"])
+    # )
+    # paramdict["dataset"]["num_labels"] = num_labels
+    # if debug_on_global:
+    #     debug(f"The number of unique labels is {num_labels}")
 
-    return PipelineOutput(
-        paramdict, tokenizer, dataset_tokenized, model, metrics_loaded
-    )
+    # tokenizer = _get_tokenizer(
+    #     paramdict["model_full_name"],
+    #     paramdict["save_dir"],
+    # )
+    tokenizer = ""
+
+    # dataset_tokenized = _get_tokenized_dataset(
+    #     dataset_plain,
+    #     tokenizer,
+    #     paramdict["dataset"]["cols_to_tokenize"],
+    #     paramdict["dataset"]["cols_to_remove"],
+    # )
+    dataset_tokenized = ""
+
+    paramdict["dataset"]["num_labels"] = 2
+    model = _get_model(paramdict["model_full_name"], paramdict["dataset"]["num_labels"])
+    # model = ""
+
+    # metrics_loaded = _get_metrics_to_load_objects(
+    #     paramdict["metrics"]["metrics_to_load"]
+    # )
+    metrics_loaded = ""
+
+    return {
+        "paramdict": paramdict,
+        "tokenizer": tokenizer,
+        "dataset_tokenized": dataset_tokenized,
+        "model": model,
+        "metrics_loaded": metrics_loaded,
+    }
+
+
+def _get_dataset(
+    name: str, configuration: str = None, save_dir: str = None
+) -> DatasetDict:
+    """
+    Downloads the dataset by calling the appropriate provider handling function.\n
+    To date only from Hugging Face.
+    """
+    return get_dataset_hf(name, configuration, save_dir)
+
+
+def _get_tokenizer(model_full_name: str, save_dir: str = None) -> AutoTokenizer:
+    """
+    Downloads the tokenizer by calling the appropriate provider handling function.\n
+    To date only from Hugging Face.
+    """
+    return get_tokenizer_hf(model_full_name, save_dir)
+
+
+def _get_metrics_to_load_objects(metrics_to_load: list) -> list[dict]:
+    """
+    Downloads metrics objects by calling the appropriate provider handling function.\n
+    To date only from Hugging Face.
+    """
+    return get_metrics_to_load_objects_hf(metrics_to_load)
+
+
+def _get_model(model_full_name: str = None, num_labels: str = None):
+    """
+    Downloads the model by calling the appropriate provider handling function.\n
+    To date only from Hugging Face.
+    """
+    return get_model_hf(model_full_name, num_labels)
 
 
 def _get_tokenized_dataset(
@@ -85,7 +150,10 @@ def _get_tokenized_dataset(
     cols_to_tokenize: list[str],
     cols_to_remove: list[str],
 ) -> DatasetDict:
-    """TODO"""
+    """
+    Returns the sanitized tokenized dataset stripped of columns not needed.\n
+    TODO save a local copy of the tokenized dataset to avoid overhead
+    """
 
     if debug_on_global:
         debug(
@@ -95,6 +163,7 @@ def _get_tokenized_dataset(
         )
 
     try:
+        # TODO save local copy of tokenized dataset
         ds_tokenized = (
             _get_raw_tokenized_dataset(dataset_plain, tokenizer, cols_to_tokenize)
             .remove_columns(cols_to_tokenize)
@@ -102,7 +171,7 @@ def _get_tokenized_dataset(
         )
         if debug_on_global:
             ds_tokenized_train_slice = ds_tokenized["train"][0]
-            print(f"{ds_tokenized_train_slice=}")
+            debug(f"{ds_tokenized_train_slice=}")
         return ds_tokenized
     except Exception as e:
         return e
@@ -111,28 +180,13 @@ def _get_tokenized_dataset(
 def _get_raw_tokenized_dataset(
     dataset_plain: DatasetDict, tokenizer: object, cols_to_tok: list[str]
 ) -> DatasetDict:
-    """Returns tokenized dataset"""
+    """Returns the raw tokenized dataset with all columns"""
 
     def _tokenize(ds):
         cols = [ds[col] for col in cols_to_tok]
         return tokenizer(*cols, truncation=True)
 
     return dataset_plain.map(_tokenize, batched=True)
-
-
-def _get_metrics_to_load_objects(metrics_to_load: list) -> list[dict]:
-    """
-    Downloads metrics objects by calling the appropriate handling function.
-    To date onyl from Hugging Face
-    """
-    # TODO implement other providers
-    return get_metrics_to_load_objects_hf(metrics_to_load)
-
-
-def _get_model(model_full_name: str = None, num_labels: str = None):
-    """TODO"""
-    # TODO implement other providers
-    return get_model_hf(model_full_name, num_labels)
 
 
 def _set_provider_env(provider: str, provider_param: dict) -> None: